[ALPS10087486] mml: update dispatch rule
[Description]
Update dispatch rule:
1. cache_mode_caps(): change NOT-SUPPORT return by
   info->mode to by info_cache->mode_caps = 0
2. add racing mode
3. merge some dc/dc2 conditions

[Test]
Display stress test

MTK-Commit-Id: d6059bfc2f5d1ce76ad596f8755d32b85d4f630c

CR-Id: ALPS10087486
Feature: mml

+++ drivers/gpu/drm/mediatek/mml/mtk-mml-core.h
@@ -378,7 +378,7 @@ struct mml_frame_info_cache {
 	u32 remain;
 	u8 dl_opp;
 	u8 dc_opp;
-	u32 mode_caps;
+	u16 mode_caps;
 };
 
 struct mml_topology_info {
